Grab Bag Decompilation (#102)
* Decomp Part Of zMusic.cpp

Progress.

* Cleanup zMusic Changes

For merging.

* Partial Decompilation Of zTaxi.cpp

9/13 functions done.

* Decompile Most Of zCameraFly.cpp; Cleanup

More progress.

* zRumble, zFX, Snd zScript Partial Decpmpilation

More general progress.

* zRumble, zFX, And zScript Partial Decpmpilation

More general progress.

* Cleanup For Merging

For progress.

* Another Cleanup For Pull Request

3rd time's the charm.

* Partial Decompilation Of iTime.cpp

Progress!

* Decompile Most Of Particle Manager

So close to 100%ing it, but yet so far.

* Start xBehaviour; Remove zScript.s

I need sleep.

* More Work On xBehaviour

Progress.

* Partial Decompilation Of Various Files

Include xDebug, zPendulum, and zTaskBox. All work in progress.

* Partial Decompilation Of zMovePoint And zLightEffect

Progress.

* Partial Decompilation Of xGroup

Progress.

* Remove Nonexistent Link Variable

It doesn't exist.

* Decompile Most Of xRMemData

So close, yet so far...

* Fully Decompile xFog.cpp

But removing the ASM file doesn't work because of the float :{

* Random Progress Dump

A bunch of stuff that works or not.

* Partial Progress On xSurface.cpp

Progress.